Search non-linearly in fromCartesian for precision
fromCartesian was off by at least 1 meter for almost every calculation.
This change causes it to search non-linearly in order close in on an
answer that is usually within 1 millimeter. However, the cost of this
change is that the number of loops needed to find the answer
dramatically increases for short distances. For example, two coordinates
that are 1 meter apart used to take 1 loop iteration to find the answer.
It now takes over 100 iterations.

This algorithm could be optimized further.